Grupper skjemaoversikt

Grupperer skjemaoversikt i kategoriar alt etter om skjemaa er ferdige eller uferdige, og slår saman ad hoc oppfylgjing og oppfylgjingskjema frå og med år 5 til kategorien «Videre oppfølging».

grupper_skjemaoversikt(d_skjemaoversikt)

Arguments

d_skjemaoversikt

Skjemaet «SkjemaOversikt» frå LTMV-databasen, som ei dataramme.

Value

Dataramma d_skjemaoversikt med to ekstra kolonner, skjema_gruppe_nr og skjema_gruppe.

Details

Tak inn skjemaet «SkjemaOversikt» frå LTMV-databasen, som kan hentast med hent_skjema("SkjemaOversikt"), og legg til to nye kolonner, skjema_gruppe_nr og skjema_gruppe.

skjema_gruppe er lik skjemanavn for alle skjema utanom oppfylgjing frå og med år 5, og ad hoc oppfylgjing. Desse vert slått saman og får verdien Videre oppfølging (år 5+ og AdHoc) i den nye kolonnen skjema_gruppe. I tillegg vil det for uferdige skjema leggjast på "(uferdig)" på slutten av verdiane i skjema_gruppe for uferdige skjema.

skjema_gruppe_nr tek utgangspunkt i kolonnen skjemarekkeflg, som er eit unikt tal for kvar skjematype, der verdien svarar til oppfylgjingsår. Ad hoc oppfylgjing har i skjemarekkeflg verdiar på 500 og noko, og avslutningsskjema verdien 999. skjema_gruppe_nr nyttar same verdiane som skjemarekkeflg, utanom oppfylgjing frå og med år 5 og ad hoc oppfylgjing, som no får verdien 99. I tillegg skiljer skjema_gruppe_nr på ferdige og uferdige skjema, i motsetnad til skjemarekkeflg, ved å leggja til 0.5 til verdien. I skjema_gruppe_nr tek altså ferdige skjema heiltalsverdiar, medan uferdige skjema har tilsvarande verdiar, men med .5 lagt til.

Dette er ein intern funksjon som i hovudsak nyttast til å laga tabelloversikt over ferdige og uferdige skjema.

Examples

if (FALSE) { # \dontrun{
library(dplyr)

d_skjemaoversikt_gruppert = hent_skjema("SkjemaOversikt") %>%
  ltmv:::grupper_skjemaoversikt()

d_skjemaoversikt_gruppert %>%
  head(20) %>%
  select(skjemanavn, skjemastatus, skjema_gruppe_nr, skjema_gruppe)
} # }