Les inn LTMV-data

Les inn LTMV-datadumpar og lagra dei som objekt, eventuelt med automatisk validering.

les_data_ltmv(
  mappe_dd,
  dato = NULL,
  maksdato = NULL,
  status = 1,
  valider = TRUE,
  omgjevnad = .GlobalEnv
)

Arguments

mappe_dd

Grunnmappe som datadumpar skal hentast frå. Må ha undermapper med datonamn (format: ÅÅÅÅ-MM-DD).

dato

Datoen for den aktuelle datadumpen (anten tekst på formatet ÅÅÅÅ-MM-DD eller eit base::Date-objekt). Vert brukt til å velja rett undermappe til mappe_dd. Viss NULL (standard), vert nyaste dato brukt.

maksdato

Dato for siste startdato, oppfølgingsdato og avslutningsdato som skal vera med i datauttrekket (same datoformat som for dato). Oppføringar som tilhøyrer seinare datoar, vert filtrerte ut frå skjemaa. Viss NULL (standard), vert det ikkje noko filtrering på datoane.

status

Type oppføringar som skal lesast. Sjå rapwhale::les_dd_oqr() for informasjon om moglege verdiar. Som standard vert berre ferdigstilte oppføringar lesne.

valider

Skal dataa validerast, dvs. sjekkast for feilverdiar før innlesing? (Standard: ja.) Omfattar testar ved bruk av kodeboka.

omgjevnad

Omgjevnad som dataobjekta skal lagrast til. Standard er den globale omgjevnaden.

Details

Alle standard datafiler til registeret vert lesne inn, eventuelt validerte, filtrerte (sjå nedanfor) og så lagra som R-objekt. Dei får namna d_full_skjemanamn, der skjemanamn er namnet på det aktuelle skjemaet (for eksempel d_full_ventfol for oppfølgingskjemaet). Som standard vert dei lagra i den globale omgjevnaden, og dei vil overskriva eventuelle eksisterande objekt med same namn.

I tillegg til alle standardskjemaa som er nemnde i kodeboka, vert datasett for mce- og patientlist-filene òg lagra.

Kodeboka for registeret vert òg lagra, med namnet kb.

Examples

if (FALSE) { # \dontrun{
# Les inn ferdigstilte skjema (nyaste datadumpar)
les_data_ltmv()
} # }